Adding typescript, sass and rearranging files
