+++ src/app/App.test.js
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import ReactDOM from "react-dom";
+import App from './App';
+
+test('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<App />, div);
+});
