+++ src/app/App.tsx
@@ -1,0 +1,14 @@
+import * as React from "react";
+import './App.scss';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        Hello world
+      </header>
+    </div>
+  );
+}
+
+export default App;
