+++ src/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.scss';
+import App from './app/App';
+
+ReactDOM.render(<App />, document.getElementById('root'));
